refactor actions for better readabiluty and efficiency
