+++ actions/getCategories.ts
